Compile and run a TreePPL program

Compile and run a TreePPL program

Usage

run_treeppl(
  source,
  data_path,
  method = "smc-bpf",
  samples = 1000,
  chains = 1,
  out_path = NULL,
  name_exe = "out",
  sample_freq = 1000
)

Arguments

source

File path to TreePPL source program.

data_path

Complete path to JSON file with input data. Created by treeppl_input().

method

Inference method to be used.

samples

The number of samples/particles during inference. Default is 1000 samples.

chains

The number of chains/sweeps. Each chain will contain the number of samples passed in samples. Default is 1 chain.

out_path

Path to JSON file outputed by TreePPL.

name_exe

Name of the executable compiled from source.

sample_freq

For each sweep/chain, sample one particle/sample every sample_freq particles/samples.

Value

A data frame with sampled values, log weights, normalized weights, and the normalizing constant for all samples.

Examples